% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/webapp.R
\name{attachHtmlToolsBrowsable}
\alias{attachHtmlToolsBrowsable}
\title{Generate a browse function}
\usage{
attachHtmlToolsBrowsable(html_complete)
}
\arguments{
\item{html_complete}{A shiny tag list}
}
\value{

}
\description{
Generate a browse function
}
\examples{
None
}
