% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/browsable.R
\name{browsable_materialise}
\alias{browsable_materialise}
\title{Browsable for Materialise CSS framework}
\usage{
browsable_materialise(...)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Browsable for Materialise CSS framework
}
\examples{
None
}
