% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opportunity_cost.R
\name{computeEconomicGain}
\alias{computeEconomicGain}
\title{Compute economic (net) gain of an option}
\usage{
computeEconomicGain(choices, chosen)
}
\arguments{
\item{choices}{A list of available choices}

\item{chosen}{A character of chosen option name}
}
\value{
Populate chosen choice with its economic net gain
}
\description{
Compute economic (net) gain of an option
}
\seealso{
compute_opportunityCost
}
