% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{convert2script}
\alias{convert2script}
\title{Convert active Rmd to R script file in a purl folder under project root}
\usage{
convert2script()
}
\value{

}
\description{
Convert active Rmd to R script file in a purl folder under project root
}
\examples{
None
}
