% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\name{convert_poly_multipoly_into_multipolyOnly}
\alias{convert_poly_multipoly_into_multipolyOnly}
\title{Convert sf dataframe all polygons into multipolygons}
\usage{
convert_poly_multipoly_into_multipolyOnly(sf_data)
}
\arguments{
\item{sf_data}{A sf dataframe}
}
\value{

}
\description{
sf with geometry of the mixture of polygons and multipolygons can not be used in plotly add_sf directly. This function makes sure all geometries are multipolygons that plotly can deal wtih.
}
\examples{
None
}
