% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/webapp.R
\name{convertHTML2RTags}
\alias{convertHTML2RTags}
\title{Convert a string of HTML tag into htmltools::tags}
\usage{
convertHTML2RTags(string)
}
\arguments{
\item{string}{A string}
}
\value{
invisible of a string of htmltools::tags format. Clipboard will always be written in the converted result for the convenience of pasting back to R script.
}
\description{
Convert a string of HTML tag into htmltools::tags
}
\examples{
None
}
