% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/webapp2.R
\name{convertLibpathRelative2root}
\alias{convertLibpathRelative2root}
\title{Convert lib path of a html file content to a path relative to project root}
\usage{
convertLibpathRelative2root(htmlfilepath)
}
\arguments{
\item{htmlfilepath}{A character of absolute path}
}
\value{

}
\description{
After conversion, the original file will have all 'lib' converted to '/lib'
}
\examples{
None
}
