% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dailyLogicChecks.R
\name{count_missing_values}
\alias{count_missing_values}
\title{Count missing values for each variable in the given list}
\usage{
count_missing_values(df, variable_names_vec = "all")
}
\arguments{
\item{df}{dataset(\code{tibble}/\code{data.frame}) object from which the
variable names are chosen.}

\item{variable_names_vec}{a vector of variable names. For each of these
variables, the function calculates the number
of missing values present in their list of
values. If you want to see missing values for
all variables (columns) in the dataset, set
\code{variable_names_vec = "all"}. Also, the
default for this argument is "all".}
}
\value{
a tibble containing the list of column (variable) names
(for the underlying dataset \code{df}) and for each of those column names,
it shows the number of values that are missing in that column. It also
shows a column that displays the percentage (of total missing values)
of values missing in each column.
}
\description{
List the number of missing values for each variable in the given list of
variables in a tidy format.
}
\examples{
count_missing_values(df = dataObj, variable_names_vec = c("a", "b"))
count_missing_values(df = dataObj, variable_names_vec = "all")
count_missing_values(df = dataObj) # uses default \code{variable_names_vec = "all"}

}
