% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dailyLogicChecks.R
\name{display_duplicates}
\alias{display_duplicates}
\title{Display duplicate surveys data}
\usage{
display_duplicates(df, uniq_identifier_col)
}
\arguments{
\item{df}{dataset(\code{tibble}/\code{data.frame}) object from which the
\code{uniq_identifier_col} is chosen.}

\item{uniq_identifier_col}{a character vector of column name(s) that uniquely
identifies the dataset. In here, \code{tidyselect}
can be used to select columns. See examples below.}
}
\value{
a tibble that displays the duplicates for the given unique identifiers
in \code{uniq_identifier_col}.
}
\description{
This function displays the duplicate surveys, given a unique identifier
(which could be a single variable, or a combination of variables). In
case, its a combination of variables (that uniquely identifies the
underlying dataset in \code{df}), then this function will show
duplicate surveys for that joint (combination) unique identifier variable.
}
\note{
If for example, \code{surveyor_id} is the unique identifier col,
and after running \code{display_duplicates} function, we find
that, there are 5 rows with \code{surveyor_id} = 122. Now,
the output tibble from this function  should be interpreted
as follows: 4 of 5 of the \code{surveyor_id} = 122 are duplicates,
1 of 5 is original.
}
\examples{

count_duplicates(df = dataObj, uniq_identifier_col = c("ID"))

count_duplicates(df = dataObj, uniq_identifier_col = c("ID", "Name"))

count_duplicates(df = dataObj, uniq_identifier_col =
tidyselect::contains("abc")) # tidyselection used to select columns.

}
