% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{download_file}
\alias{download_file}
\title{Download file from url with output path set to the root}
\usage{
download_file(url)
}
\arguments{
\item{url}{}
}
\value{

}
\description{
Download file from url with output path set to the root
}
\examples{
None
}
