% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/econ_binary.R
\name{econ_binary}
\alias{econ_binary}
\title{Estimate a Binary Response Model}
\usage{
econ_binary(
  formula,
  data,
  weights = NULL,
  subset = NULL,
  na.action = stats::na.omit,
  link = c("logit", "probit"),
  vcov = c("classical", "HC0", "HC1", "HC2", "HC3", "CR0", "CR2", "stata"),
  cluster = NULL
)
}
\arguments{
\item{formula}{A model formula.}

\item{data}{A data frame.}

\item{weights}{Optional observation weights.}

\item{subset}{Optional subset of observations.}

\item{na.action}{Function determining how missing values are handled.}

\item{link}{Binary-response link function. Either "logit" or "probit".}

\item{vcov}{Covariance estimator. Available options are "classical",
"HC0", "HC1", "HC2", "HC3", "CR0", "CR2", and "stata".}

\item{cluster}{Optional clustering variable, evaluated within data.
Required when vcov is "CR0", "CR2", or "stata".}
}
\value{
An object of class "econ_binary".
}
\description{
Fits binary-response models using logit or probit links.
}
