\name{econ_vcov}
\alias{econ_vcov}
\title{Covariance matrix for econ_lm}
\usage{econ_vcov(object, type = object$vcov_type)}
\arguments{\item{object}{An econ_lm object.}\item{type}{Covariance type.}}
\description{Computes classical, heteroskedasticity-consistent, or one-way clustered covariance matrices.}
