% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/econr-package.R
\docType{package}
\name{econr-package}
\alias{econr}
\alias{econr-package}
\title{econr: A new approach to modelling}
\description{
A collection of tools to estimate and work with models in R.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/elben10/econr}
  \item Report bugs at \url{https://github.com/elben10/econr/issues}
}

}
\author{
\strong{Maintainer}: Jakob Jul Elben \email{elbenjakobjul@gmail.com}

}
\keyword{internal}
