% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{extract_outputInputLabels}
\alias{extract_outputInputLabels}
\title{Extract input output labels from the active Rmd}
\usage{
extract_outputInputLabels()
}
\value{

}
\description{
Extract input output labels from the active Rmd
}
\examples{
None
}
