% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/browsable.R
\name{generate_browsable}
\alias{generate_browsable}
\title{Generate browsable with dependencies attachment}
\usage{
generate_browsable(main_dependencies)
}
\arguments{
\item{main_dependencies}{a taglist of html dependencies}
}
\value{
a function that attaches main_dependencies to an html objects for browsing
}
\description{
Generate browsable with dependencies attachment
}
\examples{
none
}
