% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/browsable.R
\name{generate_drakebrowsable}
\alias{generate_drakebrowsable}
\title{Generate drake object ui browsable with dependencies attachment}
\usage{
generate_drakebrowsable(main_dependencies)
}
\arguments{
\item{main_dependencies}{a taglist of html dependencies}
}
\value{
a function that attaches main_dependencies to an html objects in drake for browsing
}
\description{
Generate drake object ui browsable with dependencies attachment
}
\examples{
none
}
