% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{generate_saveFunction}
\alias{generate_saveFunction}
\title{Generate saving function for list/env attaching}
\usage{
generate_saveFunction(objectname)
}
\arguments{
\item{objectname}{A character.}
}
\value{

}
\description{
Generate saving function for list/env attaching
}
\examples{
None
}
