% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/econ_lm.R
\name{glance.econ_lm}
\alias{glance.econ_lm}
\title{Glance at an econ_lm model}
\usage{
\method{glance}{econ_lm}(x, ...)
}
\arguments{
\item{x}{econ_lm object}

\item{...}{additional arguments}
}
\description{
Glance at an econ_lm model
}
