% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{googleLink_download}
\alias{googleLink_download}
\title{Download Google drive shared link to a destination folder}
\usage{
googleLink_download(googleSharedLink, destfolder = destfolder)
}
\arguments{
\item{googleSharedLink}{A shared link}

\item{destfolder}{A destination folder path}
}
\value{

}
\description{
Download Google drive shared link to a destination folder
}
\examples{
none
}
