% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpx.R
\name{gpx_as_list}
\alias{gpx_as_list}
\title{Convert imported gpx file's xml content to a list of longitude and latitude}
\usage{
gpx_as_list(xml_gpx)
}
\value{
A list of longitude-latitude pairs
}
\description{
Convert imported gpx file's xml content to a list of longitude and latitude
}
\examples{
None
}
