% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/web_dependencies.R
\name{html_dependencies}
\alias{html_dependencies}
\title{Create a list of dependencies to choose}
\usage{
html_dependencies()
}
\value{

}
\description{
Create a list of dependencies to choose
}
\examples{
none
}
