% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/webapp.R
\name{htmlDependency_path}
\alias{htmlDependency_path}
\title{Get script path after htmltools::htmlDependency is resolved}
\usage{
htmlDependency_path(htmldep)
}
\arguments{
\item{htmldep}{An htmlDependency object}
}
\value{

}
\description{
Get script path after htmltools::htmlDependency is resolved
}
\examples{
none
}
