% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/webapp2.R
\name{htmlFile_move}
\alias{htmlFile_move}
\title{Movie html file to a another folder}
\usage{
htmlFile_move(fromPath, toPath)
}
\arguments{
\item{fromPath}{A character of absolute filepath}

\item{toPath}{A character of absolute filepath}
}
\value{
the new filepath to the html file
}
\description{
Movie html file to a another folder
}
\examples{
none
}
