% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/webapp.R
\name{initiateJson}
\alias{initiateJson}
\title{Initiate external JSON embody}
\usage{
initiateJson(web)
}
\arguments{
\item{web}{An environment initiated by Web with foldername and filename specified}
}
\value{

}
\description{
Initiate external JSON embody
}
\examples{
None
}
