% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{initiateProjectAtCurrentRmd}
\alias{initiateProjectAtCurrentRmd}
\title{Initiate a Rproj based on current Rmd directory}
\usage{
initiateProjectAtCurrentRmd()
}
\value{

}
\description{
Initiate a Rproj based on current Rmd directory
}
\examples{
none
}
