% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dailyLogicChecks.R
\name{list_common_columns}
\alias{list_common_columns}
\title{List common column names in two different data sets}
\usage{
list_common_columns(df1, df2)
}
\arguments{
\item{df1}{First dataset. This should be in either of the following formats:
\code{data.frame/tbl/tbl_df}.}

\item{df2}{Second dataset. This should be in either of the following formats:
\code{data.frame/tbl/tbl_df}.}
}
\value{
a character vector of column names that are shared by \code{df1} and
\code{df2}
}
\description{
Find common column names in 2 different data sets. E.g. In doing field surveys,
checking for variables with same name in master tracking list
& field tracking list. It's a good habit to check for such
common variable names to avoid losing necessary variables from using data.
}
\examples{
list_common_columns(df1 = dataObj1, df2 = dataObj2)

}
