% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residuals.R
\name{mod_add_residuals}
\alias{mod_add_residuals}
\title{A residuals from model to data.frame}
\usage{
mod_add_residuals(data, model, var = "resid")
}
\arguments{
\item{data}{An object of type data.frame}

\item{model}{Model of class \code{mod}.}

\item{var}{The name of residual column.}
}
\value{
A data.frame where the residuals has been added in a column.
}
\description{
A residuals from model to data.frame
}
\examples{
mod <- mtcars \%>\% mod_lm(mpg~cyl)
mod_add_residuals(mtcars, mod)
}
