% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm.R
\name{mod_lm}
\alias{mod_lm}
\title{Fit linear model}
\usage{
mod_lm(data, formula)
}
\arguments{
\item{data}{An object of type data.frame}

\item{formula}{An object of type formula}
}
\value{
Returns a list of class mod.
}
\description{
\code{mod_lm} estimates linear models by minimizing the sum of squared residuals.
}
\examples{
mod_lm(mtcars, mpg~cyl)
}
