% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_matrix.R
\name{mod_matrix}
\alias{mod_matrix}
\title{Design matrix}
\usage{
mod_matrix(data, formula)
}
\arguments{
\item{data}{An object of type data.frame}

\item{formula}{An object of type formula}
}
\value{
A tibble of the design matrix.
}
\description{
\code{mod_matrix} returns the design matrix associated to the model formula.
}
\examples{
mod_matrix(mtcars, mpg~cyl)
}
