% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_lm.R
\name{model_lm}
\alias{model_lm}
\title{Fit linear model}
\usage{
model_lm(data, formula)
}
\arguments{
\item{data}{An object of type data.frame}

\item{formula}{An object of type formula}
}
\value{
Returns a list of class mod.
}
\description{
\code{model_lm} estimates linear models by minimizing the sum of squared residuals.
}
\examples{
model_lm(mtcars, mpg~cyl)
}
