% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_matrix.R
\name{model_response}
\alias{model_response}
\title{Response vector}
\usage{
model_response(data, formula)
}
\arguments{
\item{data}{An object of type data.frame}

\item{formula}{An object of type formula}
}
\value{
A vector containing the model response.
}
\description{
\code{model_response} returns the response vector associated to the model formula.
}
\examples{
model_response(mtcars, mpg~cyl)
}
