% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/webapp2.R
\name{move2destinationFolder}
\alias{move2destinationFolder}
\title{Move built webpage file to a folder which is a sub folder of serving root folder}
\usage{
move2destinationFolder(destFolder)
}
\arguments{
\item{destFolder}{A character of absolute filepath, which is a subfolder of serving root folder}
}
\value{

}
\description{
Move built webpage file to a folder which is a sub folder of serving root folder
}
\examples{
none
}
