% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{output_path}
\alias{output_path}
\title{Generate a output path referring to the sibling folder of the project folder}
\usage{
output_path(foldername = "docs")
}
\arguments{
\item{foldername}{a character, default="docs" for github page purpose}
}
\value{

}
\description{
Generate a output path referring to the sibling folder of the project folder
}
\examples{
None
}
