% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{parse_dropboxlink}
\alias{parse_dropboxlink}
\title{Parse dropbox link}
\usage{
parse_dropboxlink(imglink)
}
\arguments{
\item{imglink}{a character of dropbox url starting with www.dropbox.com}
}
\value{
a character of dropbox link that works for <img src=...>
}
\description{
Parse dropbox link
}
\examples{
None
}
