% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initiatePily.R
\name{Pily}
\alias{Pily}
\title{Input a plotly object to create a pily instance for convenient plotly exploration}
\usage{
Pily(plotlyObject)
}
\arguments{
\item{plotlyObject}{A plotly object}
}
\value{

}
\description{
Input a plotly object to create a pily instance for convenient plotly exploration
}
