% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reactive2.R
\name{produce_serverFunctionScript}
\alias{produce_serverFunctionScript}
\title{Generate server function script from active Rmd}
\usage{
produce_serverFunctionScript()
}
\value{

}
\description{
Generate server function script from active Rmd
}
\examples{
none
}
