% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/webapp2.R
\name{save_page}
\alias{save_page}
\title{Save an htmlTag (extenstion to save_html)}
\usage{
save_page(
  htmlEl,
  outputFilepath,
  dependencies = dependencies,
  fixed_root = NULL
)
}
\arguments{
\item{htmlEl}{A htmlTag.}

\item{outputFilepath}{A character of absolute filepath}

\item{dependencies}{A html dependencies (default=NULL)}

\item{fixed_root}{A character of absolute path to the root serving folder. (default = NULL, meaning outputFilepath is at the root folder)}
}
\value{

}
\description{
Save an htmlTag (extenstion to save_html)
}
\examples{
none
}
