% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/font.R
\name{setup_chinese}
\alias{setup_chinese}
\title{Setup Chinese font for ggplot2}
\usage{
setup_chinese(need2Knit = F)
}
\arguments{
\item{need2Knit}{logical, default=F. TRUE means need to have all chunks default with fig.showtext=TRUE, fig.retina=1 for knitting Rmd purpose. If only need ggplot output to be exported as a graph file, default F is okay.}
}
\value{

}
\description{
Setup Chinese font for ggplot2
}
\examples{
None
}
