% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/econ_binary.R
\name{tidy.econ_binary}
\alias{tidy.econ_binary}
\title{Tidy a Binary Response Model}
\usage{
\method{tidy}{econ_binary}(x, conf.int = TRUE, conf.level = 0.95, ...)
}
\arguments{
\item{x}{An econ_binary object.}

\item{conf.int}{Include confidence intervals.}

\item{conf.level}{Confidence level.}

\item{...}{Additional arguments.}
}
\value{
A data frame containing coefficient estimates and inference statistics.
}
\description{
Tidy a Binary Response Model
}
