% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/econ_lm.R
\name{tidy.econ_lm}
\alias{tidy.econ_lm}
\title{Tidy an econ_lm model}
\usage{
\method{tidy}{econ_lm}(x, conf.int = TRUE, conf.level = 0.95, ...)
}
\arguments{
\item{x}{econ_lm object}

\item{conf.int}{logical; include confidence intervals}

\item{conf.level}{confidence level}

\item{...}{additional arguments}
}
\description{
Tidy an econ_lm model
}
