% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{timeoutPrevention}
\alias{timeoutPrevention}
\title{Create invisible action button automatically clicked to prevent timeout of an app.}
\usage{
timeoutPrevention(seconds = 10)
}
\arguments{
\item{seconds}{numeric. seconds to automatically click the action button}
}
\value{

}
\description{
Create invisible action button automatically clicked to prevent timeout of an app.
}
\examples{
none
}
