% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/webapp2.R
\name{turnPath_relative}
\alias{turnPath_relative}
\title{Turn absolute path to relative}
\usage{
turnPath_relative(
  absolutePath,
  absoluteRootPath,
  beginWithSlash = T,
  prefix = NULL
)
}
\arguments{
\item{absolutePath}{An absolute path to convert}

\item{absoluteRootPath}{An absolute path to define the root}

\item{beginWithSlash}{A logical (default = T), to start the relative path with /}

\item{prefix}{A character of prefix (usually . or ..) to add before /}
}
\value{
the relative path
}
\description{
Turn absolute path to relative
}
\examples{
none
}
